requirements.txt
beautifulsoup4==4.10.0
pymongo==3.12.1
selenium==3.14.0
python-dotenv==0.19.1
certifi==2021.10.8
charset-normalizer==2.0.7
chardet==3.0.4
idna==3.3
requests==2.26.0
urllib3==1.26.7